WIP: Clean up, remove unused code.
Remove htmlg-based component package for now, it's not used primarily.
Move vcomponent into its place.

Disable limit of 10 updates, it was for debugging.
