+++ component/doc.go
@@ -1,2 +1,2 @@
-// Package component contains HTML components used by Go Package Store.
+// Package component contains Vecty HTML components used by Go Package Store.
 package component
