+++ component/header.go
@@ -3,71 +3,65 @@ package component
 import (
 	"fmt"
 
-	"github.com/shurcooL/htmlg"
-	"golang.org/x/net/html"
+	"github.com/gopherjs/gopherjs/js"
+	"github.com/gopherjs/vecty"
+	"github.com/gopherjs/vecty/elem"
+	"github.com/gopherjs/vecty/event"
+	"github.com/gopherjs/vecty/prop"
+	"github.com/gopherjs/vecty/style"
+	"github.com/shurcooL/Go-Package-Store/frontend/model"
 	"golang.org/x/net/html/atom"
 )
 
-type Header struct{}
-
-func (Header) Render() []*html.Node {
-	// TODO: Make this much nicer.
-	/*
-		<div style="width: 100%; text-align: center; background-color: hsl(209, 51%, 92%);">
-			<span style="background-color: hsl(209, 51%, 88%); padding: 15px; display: inline-block;">Updates</span>
-		</div>
-	*/
-	div := &html.Node{
-		Type: html.ElementNode, Data: atom.Div.String(),
-		Attr: []html.Attribute{
-			{Key: atom.Style.String(), Val: "width: 100%; text-align: center; background-color: hsl(209, 51%, 92%);"},
-		},
-		FirstChild: &html.Node{
-			Type: html.ElementNode, Data: atom.Span.String(),
-			Attr: []html.Attribute{
-				{Key: atom.Style.String(), Val: "background-color: hsl(209, 51%, 88%); padding: 15px; display: inline-block;"},
-			},
-			FirstChild: htmlg.Text("Updates"),
-		},
-	}
-	return []*html.Node{div}
+type Header struct {
+	vecty.Core
 }
 
-// UpdatesHeader combines checkingForUpdates, noUpdates and updatesHeading
+func (*Header) Render() *vecty.HTML {
+	return elem.Div(
+		style.Width("100%"), vecty.Style("text-align", "center"), vecty.Style("background-color", "hsl(209, 51%, 92%)"),
+		elem.Span(
+			vecty.Style("background-color", "hsl(209, 51%, 88%)"), vecty.Style("padding", string(style.Px(15))), vecty.Style("display", "inline-block"),
+			vecty.Text("Updates"),
+		),
+	)
+}
+
+// updatesHeader combines checkingForUpdates, noUpdates and updatesHeading
 // into one high level component.
-type UpdatesHeader struct {
-	RPs             []*RepoPresentation
+type updatesHeader struct {
+	RPs             []*model.RepoPresentation
 	CheckingUpdates bool
 }
 
-func (u UpdatesHeader) Render() []*html.Node {
-	var ns []*html.Node
+func (u updatesHeader) Render() vecty.List {
+	var ns vecty.List
 	// Show "Checking for updates..." while still checking.
 	if u.CheckingUpdates {
-		ns = append(ns, checkingForUpdates.Render()...)
+		ns = append(ns, checkingForUpdates())
 	}
 	available, updating, supported := u.status()
 	// Show "No Updates Available" if we're done checking and there are no remaining updates.
 	if !u.CheckingUpdates && available == 0 && !updating {
-		ns = append(ns, noUpdates.Render()...)
+		ns = append(ns, noUpdates())
 	}
 	// Show number of updates available and Update All button.
-	ns = append(ns, updatesHeading{
+	ns = append(ns, &updatesHeading{
 		Available:       available,
 		Updating:        updating,
 		UpdateSupported: supported, // TODO: Fetch this value from backend once.
-	}.Render()...)
+	})
 	return ns
 }
 
 // status returns available, updating, supported updates in u.RPs.
-func (u UpdatesHeader) status() (available uint, updating bool, supported bool) {
+func (u updatesHeader) status() (available uint, updating bool, supported bool) {
 	for _, rp := range u.RPs {
 		switch rp.UpdateState {
-		case Available:
+		case model.Available:
 			available++
 			supported = rp.UpdateSupported
-		case Updating:
+		case model.Updating:
 			updating = true
 		}
 	}
@@ -77,6 +71,7 @@ func (u UpdatesHeader) status() (available uint, updating bool, supported bool) 
 // updatesHeading is a heading that displays number of updates available,
 // whether updates are installing, and an Update All button.
 type updatesHeading struct {
+	vecty.Core
 	Available uint
 	Updating  bool
 
@@ -84,86 +79,69 @@ type updatesHeading struct {
 	UpdateSupported bool
 }
 
-func (u updatesHeading) Render() []*html.Node {
+func (u *updatesHeading) Render() *vecty.HTML {
 	if u.Available == 0 && !u.Updating {
 		return nil
 	}
-	h4 := &html.Node{
-		Type: html.ElementNode, Data: atom.H4.String(),
-		Attr: []html.Attribute{
-			{Key: atom.Style.String(), Val: "text-align: left;"},
-		},
+	var status string
+	if u.Available > 0 {
+		status = fmt.Sprintf("%d Updates Available", u.Available)
 	}
 	if u.Updating {
-		h4.AppendChild(htmlg.Text("Updates Installing..."))
-	} else {
-		h4.AppendChild(htmlg.Text(fmt.Sprintf("%d Updates Available", u.Available)))
+		if status != "" {
+			status += ", "
+		}
+		status += "Installing Updates..."
 	}
-	h4.AppendChild(&html.Node{
-		Type: html.ElementNode, Data: atom.Span.String(),
-		Attr: []html.Attribute{
-			{Key: atom.Style.String(), Val: "float: right;"},
-		},
-		FirstChild: u.updateAllButton(),
-	})
-	return []*html.Node{h4}
+	return elem.Heading4(
+		vecty.Style("text-align", "left"),
+		vecty.Text(status),
+		elem.Span(
+			vecty.Style("float", "right"),
+			u.updateAllButton(),
+		),
+	)
 }
 
-func (u updatesHeading) updateAllButton() *html.Node {
+func (u *updatesHeading) updateAllButton() *vecty.HTML {
 	if !u.UpdateSupported {
-		return &html.Node{
-			Type: html.ElementNode, Data: atom.Span.String(),
-			Attr: []html.Attribute{
-				{Key: atom.Style.String(), Val: "color: gray; cursor: default;"},
-				{Key: atom.Title.String(), Val: "Updating repos is not currently supported for this source of repos."},
-			},
-			FirstChild: htmlg.Text("Update All"),
-		}
+		return elem.Span(
+			style.Color("gray"), vecty.Style("cursor", "default"),
+			vecty.Property(atom.Title.String(), "Updating repos is not currently supported for this source of repos."),
+			vecty.Text("Update All"),
+		)
 	}
 	switch {
 	case u.Available > 0:
-		return &html.Node{
-			Type: html.ElementNode, Data: atom.A.String(),
-			Attr: []html.Attribute{
-				{Key: atom.Href.String(), Val: "/api/update-all"}, // TODO: Should it be a separate endpoint or what?
-				{Key: atom.Onclick.String(), Val: "UpdateAll(event);"},
-			},
-			FirstChild: htmlg.Text("Update All"),
-		}
+		return elem.Anchor(
+			prop.Href("/api/update-all"), // TODO: Should it be a separate endpoint or what?
+			event.Click(func(e *vecty.Event) {
+				// TODO.
+				fmt.Println("UpdateAll()")
+				js.Global.Get("UpdateAll").Invoke() // TODO: Do this via action?
+			}).PreventDefault(),
+			vecty.Text("Update All"),
+		)
 	case u.Available == 0:
-		return &html.Node{
-			Type: html.ElementNode, Data: atom.Span.String(),
-			Attr: []html.Attribute{
-				{Key: atom.Style.String(), Val: "color: gray; cursor: default;"},
-			},
-			FirstChild: htmlg.Text("Update All"),
-		}
+		return elem.Span(
+			style.Color("gray"), vecty.Style("cursor", "default"),
+			vecty.Text("Update All"),
+		)
 	default:
 		panic("unreachable")
 	}
 }
 
 // InstalledUpdates is a heading for installed updates.
-var InstalledUpdates = heading{Heading: atom.H3, Text: "Installed Updates"}
+func InstalledUpdates() *vecty.HTML { return heading(elem.Heading3, "Installed Updates") }
 
-var checkingForUpdates = heading{Heading: atom.H2, Text: "Checking for updates..."}
+func checkingForUpdates() *vecty.HTML { return heading(elem.Heading2, "Checking for updates...") }
 
-var noUpdates = heading{Heading: atom.H2, Text: "No Updates Available"}
+func noUpdates() *vecty.HTML { return heading(elem.Heading2, "No Updates Available") }
 
-type heading struct {
-	Heading atom.Atom
-	Text    string
-}
-
-func (h heading) Render() []*html.Node {
-	// TODO: Make this much nicer.
-	// <{{.Heading}} style="text-align: center;">{{.Text}}</{{.Heading}}>
-	hn := &html.Node{
-		Type: html.ElementNode, Data: h.Heading.String(),
-		Attr: []html.Attribute{
-			{Key: atom.Style.String(), Val: "text-align: center;"},
-		},
-		FirstChild: htmlg.Text(h.Text),
-	}
-	return []*html.Node{hn}
+func heading(heading func(markup ...vecty.MarkupOrComponentOrHTML) *vecty.HTML, text string) *vecty.HTML {
+	return heading(
+		vecty.Style("text-align", "center"),
+		vecty.Text(text),
+	)
 }
