--- vcomponent/doc.go
@@ -1,2 +1,0 @@
-// Package vcomponent contains Vecty HTML components used by Go Package Store,
-package vcomponent
